% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSAIGE.R
\name{Run_Meta_OneSet}
\alias{Run_Meta_OneSet}
\alias{Run_MetaSAIGE}
\title{Run Meta-Analysis for a Single Set}
\usage{
Run_MetaSAIGE(
  n.cohorts,
  chr,
  gwas_path,
  info_path,
  gene_file_prefix,
  col_co,
  output_path,
  ancestry = NULL,
  trait_type = "binary",
  groupfile = NULL,
  annotation = NULL,
  mafcutoff = NULL,
  verbose = TRUE
)
}
\arguments{
\item{n.cohorts}{Number of cohorts}

\item{chr}{Chromosome number}

\item{gwas_path}{Path to GWAS summary statistics files from SAIGE. The files should be delimited by white space ' '}

\item{info_path}{Path to SNP information files from SAIGE. The files should be delimited by white space ' '}

\item{gene_file_prefix}{Prefix for gene files from SAIGE-GENE+ step3_LD_mat.R. The files should be delimited by white space ' '}

\item{col_co}{Cutoff for collapsing. Default is 10}

\item{output_path}{Output path for the results}

\item{ancestry}{Ancestry identifier. Any numbers starting from 1 could be used to identify ancestries (e.g. 1, 2, 3, 4, 5, 6, 7, 8, 9, 10)}

\item{trait_type}{Trait type. Default is 'binary'}

\item{groupfile}{Groupfile for functional annotation. Same format that is used for SAIGE-GENE+ (You can modify this file to select the variants of interest)}

\item{annotation}{Functional annotation for multiple testing to annalyze should be parsed with '_' (e.g. c('lof', 'missense_lof'))}

\item{mafcutoff}{MAF cutoff for multiple testing (e.g. c(0.01, 0.001))}

\item{verbose}{Verbose. Default is TRUE}
}
\value{
A data frame with the following columns:
- CHR: Chromosome number
- GENE: Gene name
- Pval: P-value
- Pval_0.00: P-value with 0.00 rho for SKAT-O
- Pval_0.01: P-value with 0.01 rho for SKAT-O
- Pval_0.04: P-value with 0.04 rho for SKAT-O
- Pval_0.09: P-value with 0.09 rho for SKAT-O
- Pval_0.25: P-value with 0.25 rho for SKAT-O
- Pval_0.50: P-value with 0.50 rho for SKAT-O
- Pval_1.00: P-value with 1.00 rho for SKAT-O
}
\description{
This function performs meta-analysis for a set of summary statistics from multiple cohorts.
}
\examples{
Run_MetaSAIGE(2, 1, c('gwas1.txt', 'gwas2.txt'), c('info1.txt', 'info2.txt'), c('gene1_', 'gene2_'), 10, 'output.txt', c(1, 2), TRUE, 'binary')

}
