% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaSAIGE.R
\name{Run_Meta_OneSet}
\alias{Run_Meta_OneSet}
\alias{Run_MetaSAIGE}
\title{Run Meta-Analysis for a Single Set}
\usage{
Run_MetaSAIGE(
  n.cohorts,
  chr,
  gwas_path,
  info_path,
  gene_file_prefix,
  col_co,
  output_path,
  ancestry = NULL,
  verbose = TRUE,
  trait_type = "binary"
)

Run_Meta_OneSet(
  SMat.list,
  Info.list,
  n.vec,
  IsExistSNV.vec,
  n.cohort,
  Col_Cut = 10,
  GC_cutoff = 0.05,
  r.all = c(0, 0.1^2, 0.2^2, 0.3^2, 0.5^2, 0.5, 1),
  weights.beta = c(1, 25),
  IsGet_Info_ALL = True,
  ancestry = NA,
  trait_type
)
}
\arguments{
\item{n.cohorts}{Number of cohorts}

\item{chr}{Chromosome number}

\item{gwas_path}{Path to GWAS summary statistics files from SAIGE. The files should be delimited by white space ' '}

\item{info_path}{Path to SNP information files from SAIGE. The files should be delimited by white space ' '}

\item{gene_file_prefix}{Prefix for gene files from SAIGE-GENE+ step3_LD_mat.R. The files should be delimited by white space ' '}

\item{col_co}{Cutoff for collapsing. Default is 10}

\item{output_path}{Output path for the results}

\item{ancestry}{A vector of ancestry labels.}

\item{verbose}{Verbose. Default is TRUE}

\item{trait_type}{The type of trait (binary or continuous).}

\item{SMat.list}{A list of GtG matrices from each cohort.}

\item{Info.list}{A list of dataframes with single variant information (e.g. SNP ID, p-values) from each cohort.}

\item{n.vec}{A vector of sample sizes from each cohort.}

\item{IsExistSNV.vec}{A vector of binary values indicating whether the SNP exists in each cohort.}

\item{n.cohort}{The number of cohorts.}

\item{Col_Cut}{The minor allele count cutoff for collapsing.}

\item{GC_cutoff}{The p-value cutoff for genomic control.}

\item{r.all}{A vector of values for the variance component.}

\item{weights.beta}{A vector of weights for the beta coefficients.}

\item{IsGet_Info_ALL}{A logical value indicating whether to return the Info_ALL object.}
}
\value{
A data frame with the following columns:
- CHR: Chromosome number
- GENE: Gene name
- Pval: P-value
- Pval_0.00: P-value with 0.00 rho for SKAT-O
- Pval_0.01: P-value with 0.01 rho for SKAT-O
- Pval_0.04: P-value with 0.04 rho for SKAT-O
- Pval_0.09: P-value with 0.09 rho for SKAT-O
- Pval_0.25: P-value with 0.25 rho for SKAT-O
- Pval_0.50: P-value with 0.50 rho for SKAT-O
- Pval_1.00: P-value with 1.00 rho for SKAT-O

A list of meta-analysis results.
}
\description{
This function performs meta-analysis for a set of summary statistics from multiple cohorts.

This function performs meta-analysis for a set of summary statistics from multiple cohorts.
}
\examples{
Run_MetaSAIGE(2, 1, c('gwas1.txt', 'gwas2.txt'), c('info1.txt', 'info2.txt'), c('gene1_', 'gene2_'), 10, 'output.txt', c(1, 2), TRUE, 'binary')

}
